refactor: converter 위치 변경
Related t햐o : #5
