+++ src/main/java/com/umc5th/muffler/domain/expense/dto/ExpenseConverter.java
@@ -1,4 +1,4 @@
-package com.umc5th.muffler.domain.expense.converter;
+package com.umc5th.muffler.domain.expense.dto;
 
 import com.umc5th.muffler.domain.expense.dto.CategoryDetailDto;
 import com.umc5th.muffler.domain.expense.dto.NewExpenseRequest;
@@ -15,12 +15,12 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 public class ExpenseConverter {
-    public static Expense toExpenseEntity(NewExpenseRequest dto, Member member, Category category) {
+    public static Expense toExpenseEntity(NewExpenseRequest request, Member member, Category category) {
         return Expense.builder()
-            .title(dto.getExpenseTitle())
-            .memo(dto.getExpenseMemo())
-            .date(dto.getExpenseDate())
-            .cost(dto.getExpenseCost())
+            .title(request.getExpenseTitle())
+            .memo(request.getExpenseMemo())
+            .date(request.getExpenseDate())
+            .cost(request.getExpenseCost())
             .category(category)
             .member(member)
             .build();
